backend/requirements.txt
fastapi
uvicorn
numpy
pytest
python-dotenv
